--- src/CarrinhoItem.java
@@ -1,10 +1,0 @@
-public class CarrinhoItem {
-    private Produto produto;
-    private int quantidade;
-
-    public double getPrecoTotal() {
-        return produto.getPreco() * quantidade;
-    }
-
-    // Construtor, getters e setters
-}
