--- src/Categoria.java
@@ -1,8 +1,0 @@
-import java.util.List;
-
-public class Categoria {
-    private String nome;
-    private List<Produto> produtos;
-
-    // Construtor, getters e setters
-}
