--- src/Cliente.java
@@ -1,6 +1,0 @@
-public class Cliente {
-    private String nome;
-    private Carrinho carrinho;
-
-    // Métodos para gerenciar carrinho e pagamentos
-}
