--- src/GerenciarUsuario.java
@@ -1,21 +1,0 @@
-import java.util.ArrayList;
-import java.util.List;
-
-public class GerenciarUsuario {
-    private List<Usuario> usuarios = new ArrayList<>();
-
-    public void cadastrarUsuario(Usuario usuario) {
-        usuarios.add(usuario);
-    }
-
-    public Usuario getUsuario(String nome) {
-        for (Usuario usuario : usuarios) {
-            if (usuario.getNome().equals(nome)) {
-                return usuario;
-            }
-        }
-        return null;
-    }
-
-    // Construtor, getters e setters
-}
