--- src/Pagamento.java
@@ -1,7 +1,0 @@
-public class Pagamento {
-    private String metodo;
-    private double valor;
-    private Pedido pedido;
-
-    // Construtor, getters e setters
-}
