--- src/PedidoEntregue.java
@@ -1,16 +1,0 @@
-public class PedidoEntregue extends PedidoEstado {
-    @Override
-    public void proximoEstado(Pedido pedido) {
-        // Não há próximo estado
-    }
-
-    @Override
-    public void anteriorEstado(Pedido pedido) {
-        pedido.setEstado(new PedidoEnviado());
-    }
-
-    @Override
-    public String getEstado() {
-        return "Entregue";
-    }
-}
