--- src/PedidoProcessando.java
@@ -1,16 +1,0 @@
-public class PedidoProcessando extends PedidoEstado {
-    @Override
-    public void proximoEstado(Pedido pedido) {
-        pedido.setEstado(new PedidoEnviado());
-    }
-
-    @Override
-    public void anteriorEstado(Pedido pedido) {
-        pedido.setEstado(new PedidoNovo());
-    }
-
-    @Override
-    public String getEstado() {
-        return "Processando";
-    }
-}
