--- src/Usuario.java
@@ -1,11 +1,0 @@
-public class Usuario {
-    private String nome;
-
-    public Usuario(String nome) {
-        this.nome = nome;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-}
