+++ src/main/java/com/sistema/modelo/Admin.java
@@ -1,0 +1,18 @@
+package main.java.com.sistema.modelo;
+public class Admin {
+    private Estoque estoque;
+
+    public void gerenciarEstoqueDeProdutos() {
+        // Implementação para gerenciar o estoque
+    }
+
+    public void processarEAprovarPedido(Pedido pedido) {
+        pedido.proximoEstado();
+    }
+
+    public String acompanharStatusDoPedido(Pedido pedido) {
+        return pedido.getEstado();
+    }
+
+    // Construtor, getters e setters
+}
