+++ src/main/java/com/sistema/modelo/Carrinho.java
@@ -1,0 +1,20 @@
+package main.java.com.sistema.modelo;
+import java.util.List;
+
+public class Carrinho {
+    private List<CarrinhoItem> itens;
+
+    public void adicionarProduto(Produto produto, int quantidade) {
+        // Lógica para adicionar produto
+    }
+
+    public void removerProduto(Produto produto) {
+        // Lógica para remover produto
+    }
+
+    public double getValorTotal() {
+        return itens.stream().mapToDouble(CarrinhoItem::getPrecoTotal).sum();
+    }
+
+    // Construtor, getters e setters
+}
