+++ src/main/java/com/sistema/modelo/CarrinhoItem.java
@@ -1,0 +1,11 @@
+package main.java.com.sistema.modelo;
+public class CarrinhoItem {
+    private Produto produto;
+    private int quantidade;
+
+    public double getPrecoTotal() {
+        return produto.getPreco() * quantidade;
+    }
+
+    // Construtor, getters e setters
+}
