+++ src/main/java/com/sistema/modelo/Categoria.java
@@ -1,0 +1,9 @@
+package main.java.com.sistema.modelo;
+import java.util.List;
+
+public class Categoria {
+    private String nome;
+    private List<Produto> produtos;
+
+    // Construtor, getters e setters
+}
