+++ src/main/java/com/sistema/modelo/Cliente.java
@@ -1,0 +1,7 @@
+package main.java.com.sistema.modelo;
+public class Cliente {
+    private String nome;
+    private Carrinho carrinho;
+
+    // Métodos para gerenciar carrinho e pagamentos
+}
