+++ src/main/java/com/sistema/modelo/GerenciarUsuario.java
@@ -1,0 +1,22 @@
+package main.java.com.sistema.modelo;
+import java.util.ArrayList;
+import java.util.List;
+
+public class GerenciarUsuario {
+    private List<Usuario> usuarios = new ArrayList<>();
+
+    public void cadastrarUsuario(Usuario usuario) {
+        usuarios.add(usuario);
+    }
+
+    public Usuario getUsuario(String nome) {
+        for (Usuario usuario : usuarios) {
+            if (usuario.getNome().equals(nome)) {
+                return usuario;
+            }
+        }
+        return null;
+    }
+
+    // Construtor, getters e setters
+}
