+++ src/main/java/com/sistema/modelo/Pagamento.java
@@ -1,0 +1,8 @@
+package main.java.com.sistema.modelo;
+public class Pagamento {
+    private String metodo;
+    private double valor;
+    private Pedido pedido;
+
+    // Construtor, getters e setters
+}
