+++ src/main/java/com/sistema/modelo/PedidoEnviado.java
@@ -1,0 +1,17 @@
+package main.java.com.sistema.modelo;
+public class PedidoEnviado extends PedidoEstado {
+    @Override
+    public void proximoEstado(Pedido pedido) {
+        pedido.setEstado(new PedidoEntregue());
+    }
+
+    @Override
+    public void anteriorEstado(Pedido pedido) {
+        pedido.setEstado(new PedidoProcessando());
+    }
+
+    @Override
+    public String getEstado() {
+        return "Enviado";
+    }
+}
