+++ src/main/java/com/sistema/modelo/PedidoEstado.java
@@ -1,0 +1,6 @@
+package main.java.com.sistema.modelo;
+public abstract class PedidoEstado {
+    public abstract void proximoEstado(Pedido pedido);
+    public abstract void anteriorEstado(Pedido pedido);
+    public abstract String getEstado();
+}
