+++ src/main/java/com/sistema/modelo/PedidoNovo.java
@@ -1,0 +1,17 @@
+package main.java.com.sistema.modelo;
+public class PedidoNovo extends PedidoEstado {
+    @Override
+    public void proximoEstado(Pedido pedido) {
+        pedido.setEstado(new PedidoProcessando());
+    }
+
+    @Override
+    public void anteriorEstado(Pedido pedido) {
+        // Não há estado anterior
+    }
+
+    @Override
+    public String getEstado() {
+        return "Novo";
+    }
+}
