+++ src/main/java/com/sistema/modelo/Usuario.java
@@ -1,0 +1,12 @@
+package main.java.com.sistema.modelo;
+public class Usuario {
+    private String nome;
+
+    public Usuario(String nome) {
+        this.nome = nome;
+    }
+
+    public String getNome() {
+        return nome;
+    }
+}
