--- src/tests/GerenciarUsuarioTest.java
@@ -1,31 +1,0 @@
-package tests;
-
-import static org.junit.jupiter.api.Assertions.*;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-public class GerenciarUsuarioTest {
-
-    private GerenciarUsuario gerenciarUsuario;
-    private Cliente cliente;
-
-    @BeforeEach
-    public void setUp() {
-        gerenciarUsuario = new GerenciarUsuario();
-        cliente = new Cliente("Cliente A");
-    }
-
-    @Test
-    public void testCadastrarUsuario() {
-        gerenciarUsuario.cadastrarUsuario(cliente);
-        assertEquals(1, gerenciarUsuario.getUsuarios().size());
-        assertEquals(cliente, gerenciarUsuario.getUsuarios().get(0));
-    }
-
-    @Test
-    public void testConhecerUsuario() {
-        gerenciarUsuario.cadastrarUsuario(cliente);
-        Cliente encontrado = gerenciarUsuario.conhecerUsuario("Cliente A");
-        assertEquals(cliente, encontrado);
-    }
-}
